+++ source/gear/_punch-base.js
@@ -1,0 +1,23 @@
+const { GearTemplate } = require('../classes/index.js');
+const { dealDamage, changeStagger } = require('../util/combatantUtil.js');
+
+module.exports = new GearTemplate("Punch",
+	"Strike a foe for @{damage} @{element} damage",
+	"Damage x@{critMultiplier}",
+	"Technique",
+	"Untyped",
+	0,
+	(targets, user, isCrit, adventure) => {
+		const { damage, critMultiplier, element } = module.exports;
+		let pendingDamage = user.getPower() + damage;
+		if (user.element === element) {
+			changeStagger(targets, "elementMatchFoe");
+		}
+		if (isCrit) {
+			pendingDamage *= critMultiplier;
+		}
+		return dealDamage(targets, user, pendingDamage, false, element, adventure);
+	}
+).setTargetingTags({ type: "single", team: "foe", needsLivingTargets: true })
+	.setDurability(Infinity)
+	.setDamage(0);