+++ source/gear/spear-reactive.js
@@ -2,7 +2,7 @@ const { GearTemplate, Move } = require('../classes');
 const { dealDamage, changeStagger, getNames } = require('../util/combatantUtil.js');
 
 module.exports = new GearTemplate("Reactive Spear",
-	"Strike a foe for @{damage} (x@{critMultiplier} if after foe) @{element} damage", //TODONOW needs 2 bonuses?
+	"Strike a foe for @{damage} (x@{critMultiplier} if after foe) @{element} damage",
 	"Inflict @{bonus} more Stagger",
 	"Weapon",
 	"Earth",